Projekt: 1:10

+++ 1/include/common.h
@@ -10,7 +10,7 @@ typedef struct options_t {
 	double a;
 	double b;
 	// Liczba podziałów przedziału [a,b]
-	int n;
+	double n;
 	// Dokładność
 	double e;
 	// Ścieżka pliku do zapisania resultatów
